refactor tokenize out
